Conform to undocumented Terraform file convention
The .goreleaser.yml is the standard file from Terraform
Registry. It complies with the stated requirements on
https://www.terraform.io/docs/registry/providers/publishing.html

The terraform v0.13 tool fails to download from github.com unless
the zip file has a 'v' in front of the version number. This is
commonsense. The docs and goreleaser have not caught up...